chore: convert ui-extension to typescript
