+++ typings/environment.d.ts
@@ -2,7 +2,7 @@ import { QueryOptions, MetaSys, MetaSysProps, DefaultElements, Collection } from
 import { ContentType, ContentTypeProps } from './contentType'
 import { Entry, EntryProp } from './entry'
 import { Locale, CreateLocaleProps } from './locale'
-import { UIExtensionProps, UIExtension } from './uiExtension'
+import { UIExtensionProps, UIExtension } from './generated/entities/ui-extension'
 import { Stream } from 'stream'
 import { Upload } from './upload'
 import { EditorInterface } from './generated/entities/editor-interface'
