+++ typings/space.d.ts
@@ -15,7 +15,7 @@ import { SpaceMember } from './spaceMember'
 import { Role, RoleProps } from './role'
 import { SpaceMembershipProps, SpaceMembership } from './spaceMembership'
 import { TeamSpaceMembershipProps, TeamSpaceMembership } from './teamSpaceMembership'
-import { UIExtension, UIExtensionProps } from './uiExtension'
+import { UIExtension, UIExtensionProps } from './generated/entities/ui-extension'
 import { Upload } from './upload'
 import { Stream } from 'stream'
 import { Snapshot } from './snapshot'
